requirements.txt
django==5.2.6
d20==1.1.2
pyyaml==6.0.2
